refactor runtime boot and app event wiring
