+++ src/yoyopod/runtime/boot/screens_boot.py
@@ -1,0 +1,394 @@
+"""Screen construction and registration during startup."""
+
+from __future__ import annotations
+
+from typing import TYPE_CHECKING, Any
+
+from yoyopod.coordinators import AppRuntimeState
+from yoyopod.runtime.voice import VoiceCommandExecutor, VoiceRuntimeCoordinator, VoiceSettingsResolver
+from yoyopod.ui.input import InteractionProfile
+from yoyopod.voice import VoiceSettings
+
+if TYPE_CHECKING:
+    from yoyopod.app import YoyoPodApp
+
+
+class ScreensBoot:
+    """Build screen objects and resolve initial navigation state."""
+
+    def __init__(self, app: "YoyoPodApp", *, logger: Any) -> None:
+        self.app = app
+        self.logger = logger
+
+    def setup_screens(self) -> bool:
+        """Create and register all screens."""
+        self.logger.info("Setting up screens...")
+
+        try:
+            assert self.app.display is not None
+            assert self.app.context is not None
+            assert self.app.screen_manager is not None
+            from yoyopod.ui.screens.music.now_playing import (
+                NowPlayingScreen,
+                build_now_playing_actions,
+                build_now_playing_state_provider,
+            )
+            from yoyopod.ui.screens.music.playlist import PlaylistScreen
+            from yoyopod.ui.screens.music.recent import RecentTracksScreen
+            from yoyopod.ui.screens.navigation.ask import AskScreen
+            from yoyopod.ui.screens.navigation.home import HomeScreen
+            from yoyopod.ui.screens.navigation.hub import HubScreen
+            from yoyopod.ui.screens.navigation.listen import ListenScreen
+            from yoyopod.ui.screens.navigation.menu import MenuScreen
+            from yoyopod.ui.screens.system.power import (
+                PowerScreen,
+                build_power_screen_actions,
+                build_power_screen_state_provider,
+            )
+            from yoyopod.ui.screens.voip.call_history import CallHistoryScreen
+            from yoyopod.ui.screens.voip.contact_list import ContactListScreen
+            from yoyopod.ui.screens.voip.in_call import InCallScreen
+            from yoyopod.ui.screens.voip.incoming_call import IncomingCallScreen
+            from yoyopod.ui.screens.voip.outgoing_call import OutgoingCallScreen
+            from yoyopod.ui.screens.voip.quick_call import CallScreen
+            from yoyopod.ui.screens.voip.talk_contact import TalkContactScreen
+            from yoyopod.ui.screens.voip.voice_note import (
+                VoiceNoteScreen,
+                build_voice_note_actions,
+                build_voice_note_state_provider,
+            )
+
+            display = self.app.display
+            context = self.app.context
+            screen_manager = self.app.screen_manager
+            menu_items = ["Listen", "Talk", "Ask", "Setup"]
+            self.app.hub_screen = HubScreen(
+                display,
+                context,
+                music_backend=self.app.music_backend,
+                local_music_service=self.app.local_music_service,
+                voip_manager=self.app.voip_manager,
+            )
+            self.app.menu_screen = MenuScreen(display, context, items=menu_items)
+            self.app.home_screen = HomeScreen(display, context)
+            self.app.listen_screen = ListenScreen(
+                display,
+                context,
+                music_service=self.app.local_music_service,
+            )
+            voice_cfg = (
+                self.app.config_manager.get_voice_settings()
+                if self.app.config_manager is not None
+                else None
+            )
+            self.app.voice_runtime = VoiceRuntimeCoordinator(
+                context=context,
+                settings_resolver=VoiceSettingsResolver(
+                    context=context,
+                    config_manager=self.app.config_manager,
+                    settings_provider=lambda: VoiceSettings(
+                        commands_enabled=(
+                            self.app.context.voice.commands_enabled
+                            if self.app.context is not None
+                            else True
+                        ),
+                        ai_requests_enabled=(
+                            self.app.context.voice.ai_requests_enabled
+                            if self.app.context is not None
+                            else True
+                        ),
+                        screen_read_enabled=(
+                            self.app.context.voice.screen_read_enabled
+                            if self.app.context is not None
+                            else False
+                        ),
+                        stt_enabled=(
+                            self.app.context.voice.stt_enabled
+                            if self.app.context is not None
+                            else True
+                        ),
+                        tts_enabled=(
+                            self.app.context.voice.tts_enabled
+                            if self.app.context is not None
+                            else True
+                        ),
+                        mic_muted=(
+                            self.app.context.voice.mic_muted
+                            if self.app.context is not None
+                            else False
+                        ),
+                        output_volume=self.app.get_output_volume()
+                        or (
+                            self.app.context.voice.output_volume
+                            if self.app.context is not None
+                            else 50
+                        ),
+                        stt_backend=(
+                            voice_cfg.assistant.stt_backend if voice_cfg is not None else "vosk"
+                        ),
+                        tts_backend=(
+                            voice_cfg.assistant.tts_backend
+                            if voice_cfg is not None
+                            else "espeak-ng"
+                        ),
+                        vosk_model_path=(
+                            voice_cfg.assistant.vosk_model_path
+                            if voice_cfg is not None
+                            else "models/vosk-model-small-en-us"
+                        ),
+                        vosk_model_keep_loaded=(
+                            voice_cfg.assistant.vosk_model_keep_loaded
+                            if voice_cfg is not None
+                            else True
+                        ),
+                        speaker_device_id=(
+                            self.app.context.voice.speaker_device_id
+                            if self.app.context is not None
+                            and self.app.context.voice.speaker_device_id is not None
+                            else (
+                                voice_cfg.audio.speaker_device_id.strip() or None
+                                if voice_cfg is not None
+                                else None
+                            )
+                        ),
+                        capture_device_id=(
+                            self.app.context.voice.capture_device_id
+                            if self.app.context is not None
+                            and self.app.context.voice.capture_device_id is not None
+                            else (
+                                voice_cfg.audio.capture_device_id.strip() or None
+                                if voice_cfg is not None
+                                else None
+                            )
+                        ),
+                        sample_rate_hz=(
+                            voice_cfg.assistant.sample_rate_hz if voice_cfg is not None else 16000
+                        ),
+                        record_seconds=(
+                            voice_cfg.assistant.record_seconds if voice_cfg is not None else 4
+                        ),
+                        tts_rate_wpm=(
+                            voice_cfg.assistant.tts_rate_wpm if voice_cfg is not None else 155
+                        ),
+                        tts_voice=voice_cfg.assistant.tts_voice if voice_cfg is not None else "en",
+                    ),
+                ),
+                command_executor=VoiceCommandExecutor(
+                    context=context,
+                    config_manager=self.app.config_manager,
+                    people_directory=self.app.people_directory,
+                    voip_manager=self.app.voip_manager,
+                    volume_up_action=self.app.volume_up,
+                    volume_down_action=self.app.volume_down,
+                    mute_action=(
+                        self.app.voip_manager.mute if self.app.voip_manager is not None else None
+                    ),
+                    unmute_action=(
+                        self.app.voip_manager.unmute if self.app.voip_manager is not None else None
+                    ),
+                    play_music_action=(
+                        self.app.local_music_service.shuffle_all
+                        if self.app.local_music_service is not None
+                        else None
+                    ),
+                ),
+            )
+            self.app.ask_screen = AskScreen(
+                display,
+                context,
+                config_manager=self.app.config_manager,
+                people_directory=self.app.people_directory,
+                voip_manager=self.app.voip_manager,
+                volume_up_action=self.app.volume_up,
+                volume_down_action=self.app.volume_down,
+                mute_action=(
+                    self.app.voip_manager.mute if self.app.voip_manager is not None else None
+                ),
+                unmute_action=(
+                    self.app.voip_manager.unmute if self.app.voip_manager is not None else None
+                ),
+                play_music_action=(
+                    self.app.local_music_service.shuffle_all
+                    if self.app.local_music_service is not None
+                    else None
+                ),
+                voice_runtime=self.app.voice_runtime,
+            )
+            self.app.power_screen = PowerScreen(
+                display,
+                context,
+                state_provider=build_power_screen_state_provider(
+                    power_manager=self.app.power_manager,
+                    network_manager=self.app.network_manager,
+                    status_provider=self.app.get_status,
+                    playback_device_options_provider=(
+                        self.app.audio_device_catalog.playback_devices
+                        if self.app.audio_device_catalog is not None
+                        else None
+                    ),
+                    capture_device_options_provider=(
+                        self.app.audio_device_catalog.capture_devices
+                        if self.app.audio_device_catalog is not None
+                        else None
+                    ),
+                ),
+                actions=build_power_screen_actions(
+                    network_manager=self.app.network_manager,
+                    refresh_voice_device_options_action=(
+                        self.app.audio_device_catalog.refresh_async
+                        if self.app.audio_device_catalog is not None
+                        else None
+                    ),
+                    persist_speaker_device_action=(
+                        self.app.config_manager.set_voice_speaker_device_id
+                        if self.app.config_manager is not None
+                        else None
+                    ),
+                    persist_capture_device_action=(
+                        self.app.config_manager.set_voice_capture_device_id
+                        if self.app.config_manager is not None
+                        else None
+                    ),
+                    volume_up_action=self.app.volume_up,
+                    volume_down_action=self.app.volume_down,
+                    mute_action=(
+                        self.app.voip_manager.mute if self.app.voip_manager is not None else None
+                    ),
+                    unmute_action=(
+                        self.app.voip_manager.unmute if self.app.voip_manager is not None else None
+                    ),
+                ),
+            )
+            self.app.now_playing_screen = NowPlayingScreen(
+                display,
+                context,
+                state_provider=build_now_playing_state_provider(
+                    context=context,
+                    music_backend=self.app.music_backend,
+                ),
+                actions=build_now_playing_actions(
+                    context=context,
+                    music_backend=self.app.music_backend,
+                ),
+            )
+            self.app.playlist_screen = PlaylistScreen(
+                display,
+                context,
+                music_service=self.app.local_music_service,
+            )
+            self.app.recent_tracks_screen = RecentTracksScreen(
+                display,
+                context,
+                music_service=self.app.local_music_service,
+            )
+            self.app.call_screen = CallScreen(
+                display,
+                context,
+                voip_manager=self.app.voip_manager,
+                people_directory=self.app.people_directory,
+                call_history_store=self.app.call_history_store,
+            )
+            self.app.call_history_screen = CallHistoryScreen(
+                display,
+                context,
+                voip_manager=self.app.voip_manager,
+                call_history_store=self.app.call_history_store,
+            )
+            self.app.talk_contact_screen = TalkContactScreen(
+                display,
+                context,
+                voip_manager=self.app.voip_manager,
+            )
+            self.app.contact_list_screen = ContactListScreen(
+                display,
+                context,
+                voip_manager=self.app.voip_manager,
+                people_directory=self.app.people_directory,
+            )
+            self.app.voice_note_screen = VoiceNoteScreen(
+                display,
+                context,
+                state_provider=build_voice_note_state_provider(
+                    context=context,
+                    voip_manager=self.app.voip_manager,
+                ),
+                actions=build_voice_note_actions(voip_manager=self.app.voip_manager),
+            )
+            self.app.incoming_call_screen = IncomingCallScreen(
+                display,
+                context,
+                voip_manager=self.app.voip_manager,
+                caller_address="",
+                caller_name="Unknown",
+            )
+            self.app.outgoing_call_screen = OutgoingCallScreen(
+                display,
+                context,
+                voip_manager=self.app.voip_manager,
+                callee_address="",
+                callee_name="Unknown",
+            )
+            self.app.in_call_screen = InCallScreen(
+                display,
+                context,
+                voip_manager=self.app.voip_manager,
+            )
+
+            screen_manager.register_screen("hub", self.app.hub_screen)
+            screen_manager.register_screen("home", self.app.home_screen)
+            screen_manager.register_screen("menu", self.app.menu_screen)
+            screen_manager.register_screen("listen", self.app.listen_screen)
+            screen_manager.register_screen("ask", self.app.ask_screen)
+            screen_manager.register_screen("power", self.app.power_screen)
+            screen_manager.register_screen("now_playing", self.app.now_playing_screen)
+            screen_manager.register_screen("playlists", self.app.playlist_screen)
+            screen_manager.register_screen("recent_tracks", self.app.recent_tracks_screen)
+            screen_manager.register_screen("call", self.app.call_screen)
+            screen_manager.register_screen("talk_contact", self.app.talk_contact_screen)
+            screen_manager.register_screen("call_history", self.app.call_history_screen)
+            screen_manager.register_screen("contacts", self.app.contact_list_screen)
+            screen_manager.register_screen("voice_note", self.app.voice_note_screen)
+            screen_manager.register_screen("incoming_call", self.app.incoming_call_screen)
+            screen_manager.register_screen("outgoing_call", self.app.outgoing_call_screen)
+            screen_manager.register_screen("in_call", self.app.in_call_screen)
+            self.logger.info("    - Whisplay root: hub")
+
+            self.logger.info("  All screens registered")
+            self.logger.info("    - Listen flow: listen, playlists, recent_tracks, now_playing")
+            self.logger.info("    - Ask flow: ask")
+            self.logger.info("    - Power screen: power")
+            self.logger.info(
+                "    - VoIP screens: call, talk_contact, call_history, contacts, voice_note, incoming_call, outgoing_call, in_call"
+            )
+            self.logger.info("    - Navigation: home, menu")
+
+            initial_screen = self.get_initial_screen_name()
+            screen_manager.push_screen(initial_screen)
+            self.app._ui_state = self.get_initial_ui_state()
+            self.logger.info(f"  Initial route resolved to {initial_screen}")
+            self.logger.info(f"  Initial screen confirmed as {initial_screen}")
+            self.logger.info("  Initial screen set")
+            return True
+        except Exception:
+            self.logger.exception("Failed to setup screens")
+            return False
+
+    def get_interaction_profile(self) -> InteractionProfile:
+        """Return the active hardware interaction profile."""
+        if self.app.input_manager is not None:
+            return self.app.input_manager.interaction_profile
+        if self.app.context is not None:
+            return self.app.context.interaction_profile
+        return InteractionProfile.STANDARD
+
+    def get_initial_screen_name(self) -> str:
+        """Return the root screen for the active interaction profile."""
+        if self.get_interaction_profile() == InteractionProfile.ONE_BUTTON:
+            return "hub"
+        return "menu"
+
+    def get_initial_ui_state(self) -> AppRuntimeState:
+        """Return the base runtime state for the active interaction profile."""
+        if self.get_interaction_profile() == InteractionProfile.ONE_BUTTON:
+            return AppRuntimeState.HUB
+        return AppRuntimeState.MENU
